Ripped the bandage off: we once again have a single SCP implementation.  Didn't keep the pseudo factory as the Optic impl will need unit test updates.  We can just deploy implementations using branches.
